Categoried in Folder and made changes to the CmakeLists
