--- images.qrc
@@ -1,1 +1,0 @@
-<RCC/>
