+++ src/gui/images.qrc
@@ -1,0 +1,1 @@
+<RCC/>
